feat(pie-monorepo-utils): DSW-3604 add utility functions to detect changes and raise PRs
