+++ packages/tools/pie-monorepo-utils/eslint-plugin-update/create-eslint-plugin-update-pr.js
@@ -1,54 +1,18 @@
 /* eslint-disable camelcase */
-const BRANCH_PREFIX = 'update-eslint-snacks-plugin-'; // TODO use reference to avoid hardcoding this in multiple places
-
-function formatPrBody (diff) {
-    const sections = [];
-
-    if (diff.added.length > 0) {
-        const items = diff.added
-            .map(({ snacksComponent, piePackage, status }) => `- \`${snacksComponent}\` → \`${piePackage}\` (${status})`)
-            .join('\n');
-        sections.push(`## Added\n\n${items}`);
-    }
-
-    if (diff.removed.length > 0) {
-        const items = diff.removed
-            .map(({ snacksComponent, piePackage }) => `- \`${snacksComponent}\` (was \`${piePackage}\`)`)
-            .join('\n');
-        sections.push(`## Removed\n\n${items}`);
-    }
-
-    if (diff.statusChanged.length > 0) {
-        const items = diff.statusChanged
-            .map(({
-                snacksComponent, piePackage, from, to,
-            }) => `- \`${snacksComponent}\` → \`${piePackage}\`: ${from} → **${to}**`)
-            .join('\n');
-        sections.push(`## Status Changes\n\n${items}`);
-    }
-
-    if (diff.snacksChanged.length > 0) {
-        const items = diff.snacksChanged
-            .map(({
-                snacksComponent, from, to, status,
-            }) => `- \`${snacksComponent}\` (${status}): \`${from}\` → \`${to}\``)
-            .join('\n');
-        sections.push(`## Package Changes\n\n${items}`);
-    }
-
-    return sections.join('\n\n');
-}
+const { BRANCH_PREFIX, formatPrBody } = require('./shared');
 
 module.exports = async ({ github, branchName, changesSummary }) => {
     const [owner, repo] = process.env.GITHUB_REPOSITORY.split('/');
 
     if (!branchName.startsWith(BRANCH_PREFIX)) return null;
 
+    // List open PRs to find stale ones
     const { data: openPRs } = await github.rest.pulls.list({
         owner,
         repo,
         state: 'open',
     });
+    const stalePRs = openPRs.filter((pr) => pr.head.ref.startsWith(BRANCH_PREFIX));
 
     // Create new PR
     const diff = JSON.parse(changesSummary);
@@ -66,7 +30,6 @@ module.exports = async ({ github, branchName, changesSummary }) => {
     console.info(`✅ PR created: ${html_url}`);
 
     // Remove stale PRs
-    const stalePRs = openPRs.filter((pr) => pr.head.ref.startsWith(BRANCH_PREFIX));
     await Promise.all(stalePRs.map(async (pr) => {
         console.info(`Closing stale PR #${pr.number}: ${pr.title}`);
         await github.rest.pulls.update({
